Move MatomoApiClient to separate app
