+++ integreat_cms/locale/de/LC_MESSAGES/django.po
@@ -6351,22 +6351,6 @@ msgstr ""
 msgid "Unknown"
 msgstr "Unbekannt"
 
-#: cms/utils/matomo_api_manager.py
-msgid "Total Accesses"
-msgstr "Alle Zugriffe"
-
-#: cms/utils/matomo_api_manager.py
-msgid "Offline Accesses"
-msgstr "Offline Zugriffe"
-
-#: cms/utils/matomo_api_manager.py
-msgid "WebApp Accesses"
-msgstr "WebApp Zugriffe"
-
-#: cms/utils/matomo_api_manager.py
-msgid "CW"
-msgstr "KW"
-
 #: cms/utils/pdf_utils.py
 msgid "No valid pages selected for PDF generation."
 msgstr ""
@@ -7619,6 +7603,22 @@ msgstr "{} \"{}\" wurde erfolgreich übersetzt"
 msgid "{} \"{}\" could not be automatically translated."
 msgstr "{} \"{}\" konnte nicht automatisch übersetzt werden."
 
+#: matomo_api/matomo_api_client.py
+msgid "Total Accesses"
+msgstr "Alle Zugriffe"
+
+#: matomo_api/matomo_api_client.py
+msgid "Offline Accesses"
+msgstr "Offline Zugriffe"
+
+#: matomo_api/matomo_api_client.py
+msgid "WebApp Accesses"
+msgstr "WebApp Zugriffe"
+
+#: matomo_api/matomo_api_client.py
+msgid "CW"
+msgstr "KW"
+
 #: summ_ai_api/utils.py
 msgid "No German translation could be found for {} \"{}\"."
 msgstr "Es konnte keine Deutsche Übersetzung für {} \"{}\" gefunden werden."
@@ -7732,6 +7732,9 @@ msgstr ""
 
 #~ msgid "Select days and time"
 #~ msgstr "Tage und Zeit auswählen"
+
+#~ msgid "Offline Downloads"
+#~ msgstr "Offline Downloads"
 
 #~ msgid "Delete this organization"
 #~ msgstr "Diese Organisation löschen"
