+++ integreat_cms/locale/nl/LC_MESSAGES/django.po
@@ -1659,7 +1659,7 @@ msgid "Language"
 msgstr "Taal"
 
 #: cms/forms/feedback/region_feedback_filter_form.py:20
-#: cms/utils/matomo_api_manager.py:212 cms/utils/matomo_api_manager.py:358
+#: cms/utils/matomo_api_client.py:212 cms/utils/matomo_api_client.py:358
 msgid "All languages"
 msgstr "Alle talen"
 
@@ -6003,11 +6003,11 @@ msgstr ""
 msgid "Unknown"
 msgstr "Onbekend"
 
-#: cms/utils/matomo_api_manager.py:350
+#: cms/utils/matomo_api_client.py:350
 msgid "Offline Downloads"
 msgstr "Offline downloads"
 
-#: cms/utils/matomo_api_manager.py:407 cms/utils/matomo_api_manager.py:412
+#: cms/utils/matomo_api_client.py:407 cms/utils/matomo_api_client.py:412
 msgid "CW"
 msgstr "CW"
